+++ src/app/_components/Chart/index.tsx
@@ -1,0 +1,23 @@
+import { ChartContainer, ChartStyle } from "./Container";
+import {
+  ChartContext,
+  type ChartConfig,
+  useChart,
+  getPayloadConfigFromPayload,
+} from "./Context";
+import { ChartLegend, ChartLegendContent } from "./Legend";
+import { ChartTooltip, ChartTooltipContent } from "./Tooltip";
+
+export {
+  ChartContainer,
+  ChartStyle,
+  ChartContext,
+  useChart,
+  getPayloadConfigFromPayload,
+  ChartLegend,
+  ChartLegendContent,
+  ChartTooltip,
+  ChartTooltipContent,
+};
+
+export type { ChartConfig };
